ENG-8525 Create upgrader test for Deploy Installer

+++ src/main/kotlin/ai/digital/integration/server/deploy/internals/cluster/operator/AwsEksHelper.kt
@@ -22,8 +22,7 @@ open class AwsEksHelper(project: Project) : OperatorHelper(project) {
 
         updateKubeConfig()
 
-        val kubeContextInfo = getKubectlHelper().getCurrentContextInfo(skip = true)
-        updateInfrastructure(kubeContextInfo)
+        updateInfrastructure()
 
         if (getStorageClass() == "aws-efs") createStorageClass(getStorageClass())
 
@@ -355,15 +354,17 @@ open class AwsEksHelper(project: Project) : OperatorHelper(project) {
 
         undeployCluster()
 
-        project.logger.lifecycle("Delete iamserviceaccount for CSI driver.")
-        deleteIAMRoleForCSIDriver(getProvider())
+        if (awsEksProvider.destroyClusterOnShutdown.get()) {
+            project.logger.lifecycle("Delete iamserviceaccount for CSI driver.")
+            deleteIAMRoleForCSIDriver(getProvider())
 
-        project.logger.lifecycle("Delete cluster and ssh key")
-        deleteSshKey(awsEksProvider)
-        deleteCluster(awsEksProvider)
+            project.logger.lifecycle("Delete cluster and ssh key")
+            deleteSshKey(awsEksProvider)
+            deleteCluster(awsEksProvider)
 
-        project.logger.lifecycle("Delete current context")
-        getKubectlHelper().deleteCurrentContext()
+            project.logger.lifecycle("Delete current context")
+            getKubectlHelper().deleteCurrentContext()
+        }
     }
 
     private fun deleteIAMRoleForCSIDriver(awsEksProvider: AwsEksProvider) {
@@ -431,7 +432,8 @@ open class AwsEksHelper(project: Project) : OperatorHelper(project) {
         return getProvider().storageClass.getOrElse("aws-efs")
     }
 
-    private fun updateInfrastructure(infraInfo: InfrastructureInfo) {
+    private fun updateInfrastructure() {
+        val infraInfo = getCurrentContextInfo()
         val file = File(getProviderHomeDir(), OPERATOR_INFRASTRUCTURE_PATH)
         val awsEksProvider: AwsEksProvider = getProvider()
         val pairs = mutableMapOf<String, Any>(
@@ -460,5 +462,7 @@ open class AwsEksHelper(project: Project) : OperatorHelper(project) {
     override fun getMqStorageClass(): String {
         return ("gp2")
     }
+
+    override fun getCurrentContextInfo() = getKubectlHelper().getCurrentContextInfo(skip = true)
 
 }
